Reorganise some of the sources, start using sub directories with CMake
