+++ renderer/Oculus/AnnOculusInterfaceHelper.hpp
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include "systemMacro.h"
+#include <systemMacro.h>
 
 //Oculus VR API
 #include <OVR_CAPI.h>
