+++ renderer/Oculus/AnnOgreOculusRenderer.cpp
@@ -1,6 +1,7 @@
 // This is an open source non-commercial project. Dear PVS-Studio, please check it.
 // PVS-Studio Static Code Analyzer for C, C++ and C#: http://www.viva64.com
 #include "stdafx.h"
+#include "AnnOculusTouchController.hpp"
 #ifdef _WIN32
 
 #include <string>
@@ -509,32 +510,6 @@ void AnnOgreOculusRenderer::updateTouchControllers()
 
 		handController->updateVisibility();
 	}
-}
-
-void AnnOculusTouchController::rumbleStart(float factor)
-{
-	ovr_SetControllerVibration(currentSession, myControllerType, 0, factor);
-}
-
-void AnnOculusTouchController::rumbleStop()
-{
-	ovr_SetControllerVibration(currentSession, myControllerType, 0, 0);
-}
-
-AnnOculusTouchController::AnnOculusTouchController(ovrSession session,
-												   Ogre::SceneNode* handNode,
-												   AnnHandControllerID controllerID,
-												   AnnHandControllerSide controllerSide) :
- AnnHandController("Oculus Touch", handNode, controllerID, controllerSide),
- currentSession(session)
-{
-	if(side == leftHandController)
-		myControllerType = ovrControllerType_LTouch;
-	else if(side == rightHandController)
-		myControllerType = ovrControllerType_RTouch;
-
-	capabilites = RotationalTracking | PositionalTracking | AngularAccelerationTracking
-		| LinearAccelerationTracking | ButtonInputs | AnalogInputs | HapticFeedback | DiscreteHandGestures;
 }
 
 Annwvyn::AnnOgreVRRenderer* AnnRendererBootstrap_Oculus(const std::string& appName)
