+++ tests/src/Ann3DTextPlaneTest.cpp
